Remove obsolete `--cask` flag from `brew` instructions
Thank you for building and maintaining this useful tool.

I was setting it up on a new machine and noticed the Homebrew
instructions were out of date.

I don't remember when exactly, but Homebrew has integrated
Casks, so the `--cask` flag is no longer needed.

As a matter of fact, the Casks website is no longer online either.

Hope this helps.